src/dapp-test/constantinople.sol: fix interface

+++ src/dapp-tests/constantinople.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.5.2;
 import "ds-test/test.sol";
 
 contract DeadCode{
-    function dummy() external returns (uint32);
+    function dummy() external returns (uint256);
 }
 
 contract ConstantinopleTests is DSTest {
@@ -118,7 +118,7 @@ contract ConstantinopleTests is DSTest {
           a := create2(0, top, 25, salt)
         }
 
-        assertEq(uint256(DeadCode(a).dummy()), 0xdeadbeef);
+        assertEq(DeadCode(a).dummy(), 0xdeadbeef);
     }
     // TODO: test some SELFDESTRUCT properties of CREATE2
     // TODO: test EXTCODEHASH on self-destructed contract
